requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-google-genai
faiss-cpu
python-dotenv
pypdf
google-generativeai